Create Financial_Feature_Engineering.R
Work in progress. At this point the script allows you to download tickers from yahoo finance, create a matrix with several financial indicators and then plot the forecasts from an ARIMA model. Note that the difficulty within the script lied in properly applying the functions to the correct group of tickers.